requirements_new.txt
Flask==3.0.0
pandas==2.1.3
gunicorn==21.2.0
openpyxl==3.1.2
Werkzeug==3.0.1 
pytz==2023.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0

